requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
Pillow
joblib
opencv-python-headless